requirements.txt
python-telegram-bot[job-queue]>=20.7,<21.0
requests>=2.20,<3.0